backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlmodel>=0.0.16",
    "psycopg2>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "better-exceptions>=0.3.3",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1",  # Added here
]